Lots of changes to locations of classes and files
This changes the location of various files and classes.

For example, the "sdb" module now has all the "infrastructure" classes
in it, composed of various files like "sdb/command.py", "sdb/walker.py",
etc.

Additionally, the "sdb.commands" module now has all the "generic"
commands, like "echo", "cast, etc.

Lastly, the ZFS specific commands are in the new "sdb.commands.zfs"
module.
