+++ sdb/commands/linux_list.py
@@ -1,0 +1,58 @@
+#
+# Copyright 2019 Delphix
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+import argparse
+from typing import Iterable
+
+import drgn
+import sdb
+
+
+class LinuxList(sdb.Walker):
+    """ walk linux list_head """
+
+    cmdName = "linux_list"
+    inputType = "struct list_head *"
+
+    def __init__(self, prog: drgn.Program, args: str = "") -> None:
+        super().__init__(prog, args)
+        parser = argparse.ArgumentParser(description="walk a linux list")
+        parser.add_argument(
+            "offset",
+            default=0,
+            type=int,
+            nargs="?",
+            help="offset of list_head in structure",
+        )
+
+        try:
+            self.args = parser.parse_args(args.split())
+        except SystemExit:
+            #
+            # When the "-h" option is passed, argparse will throw this
+            # exception after the help message is printed. This would
+            # result in the SDB session exiting, which is not what we
+            # want to occur. Thus, we ignore this exception.
+            #
+            pass
+
+    def walk(self, input: drgn.Object) -> Iterable[drgn.Object]:
+        node = input.next
+        while node != input:
+            yield drgn.Object(
+                self.prog, type="void *", value=(int(node) - self.args.offset)
+            )
+            node = node.next
